% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\docType{data}
\name{A}
\alias{A}
\title{Sparse Matrix A}
\format{A sparse matrix containing 15 columns and 3000 observations}
\usage{
data(A)
}
\description{
Data for implementing the example given for CRASSMAT.
}
\author{
Nick Kunz <\email{nick.kunz@columbia.edu}>
}
\keyword{data}
\keyword{matrix}
\keyword{sparse}
