% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRASSMAT.R
\name{crassmat}
\alias{crassmat}
\title{CRASSMAT}
\usage{
crassmat(data, sample_thres, conditional)
}
\arguments{
\item{data}{either a sparse matrix, a complete matrix, or a matrix containing missing values}

\item{sample_thres}{a non-negative decimal specifying the percentage sampled out of the matrix from observed values}

\item{conditional}{a non-negative integer specifying the number of observed values in matrix to remain per row}
}
\value{
The returned object is a matrix with x number of values sampled without replacement based on the specified sampling threshold and conditional set by the user.
}
\description{
This package includes CRASSMAT, a function for conditional random sampling observations in sparse matrices. CRASSMAT takes a given sparse matrix Aij and samples out a single jth value on the condition that the number of jth values within the ith observation is greater than the specified conditional (minimum number of values to remain per ith observation). This process repeats itself until the specified sampling threshold is met. Although CRASSMAT was designed for applications with sparse matrices, it can also be used with complete matrices or those with missing values. This package is useful for applications where sampling out observed values in a matrix is important for testing the predictive accuracy of data imputation methods, such as matrix factorization or singular value decomposition (SVD).
}
\examples{
\dontrun{
A_train <- crassmat(data = A,            # matrix to train / test split
                    sample_thres = 0.20, # remove 20\% of observed values
                    conditional = 1)     # keep > 1 observed value per row
}
}
\author{
Nick Kunz <nick.kunz@columbia.edu>
}
\keyword{conditional}
\keyword{imputation}
\keyword{matrices}
\keyword{matrix}
\keyword{random}
\keyword{sampling}
\keyword{sparse}
