% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRASSMAT.R
\name{crassmat}
\alias{crassmat}
\title{CRASSMAT}
\usage{
crassmat(sparse_matrix, test_split_thres, conditional)
}
\arguments{
\item{sparse_matrix}{data}

\item{test_split_thres}{percent sampled out of matrix from observed values}

\item{conditional}{number of observed values in matrix to remain per row}
}
\value{
sampled matrix
}
\description{
This package includes CRASSMAT, a function for conditional random sampling out observations in sparse matrices. CRASSMAT takes a given sparse matrix Aij and samples out a single jth value on the condition that the number of jth values within the ith observation is greater than the specified conditional (minimum number of values to remain per ith observation). This process repeats itself until the specified sampling threshold is met. Although CRASSMAT was designed for applications with sparse matrices, it can also be applied to complete matrices or those with missing values. This package is useful for applications where sampling out observed values in a matrix is important for testing the predictive accuracy of data imputation methods, such as matrix factorization or singular value decomposition (SVD).
}
\examples{
## crassmat example
A_training <- crassmat(sparse_matrix = A,        ## matrix
                       test_split_thres = 0.20,  ## percent sampled out of matrix from observed values
                       conditional = 1)          ## number of observed values in matrix to remain per row

}
\keyword{conditional}
\keyword{factorization,}
\keyword{matrix}
\keyword{matrix,}
\keyword{random}
\keyword{sampling}
\keyword{sampling,}
\keyword{sparse}
\keyword{svd}
